+++ crates/idxdb-store/src/ts/schema.ts
@@ -304,12 +304,7 @@ export class MidenDatabase {
       [Table.StateSync]: indexes("id"),
       [Table.BlockHeaders]: indexes("blockNum", "hasClientNotes"),
       [Table.PartialBlockchainNodes]: indexes("id"),
-      [Table.Tags]: indexes(
-        "id++",
-        "tag",
-        "source_note_id",
-        "source_account_id"
-      ),
+      [Table.Tags]: indexes("id++", "tag", "sourceNoteId", "sourceAccountId"),
       [Table.ForeignAccountCode]: indexes("accountId"),
       [Table.Settings]: indexes("key"),
       [Table.TrackedAccounts]: indexes("&id"),