+++ crates/rust-client/src/note/note_screener.rs
@@ -4,8 +4,7 @@ use alloc::vec::Vec;
 
 use async_trait::async_trait;
 use miden_protocol::account::{Account, AccountId};
-use miden_protocol::errors::{AccountError, AssetError};
-use miden_protocol::note::{Note, NoteId};
+use miden_protocol::note::Note;
 use miden_standards::account::interface::{AccountInterface, AccountInterfaceExt};
 use miden_standards::note::NoteConsumptionStatus;
 use miden_tx::auth::TransactionAuthenticator;
@@ -185,8 +184,6 @@ where
 /// Error when screening notes to check relevance to a client.
 #[derive(Debug, Error)]
 pub enum NoteScreenerError {
-    #[error("error while processing note inputs")]
-    InvalidNoteInputsError(#[from] InvalidNoteInputsError),
     #[error("account data wasn't found for account id {0}")]
     AccountDataNotFound(AccountId),
     #[error("error while fetching data from the store")]
@@ -195,16 +192,4 @@ pub enum NoteScreenerError {
     NoteCheckerError(#[from] NoteCheckerError),
     #[error("error while building transaction request")]
     TransactionRequestError(#[from] TransactionRequestError),
-}
-
-#[derive(Debug, Error)]
-pub enum InvalidNoteInputsError {
-    #[error("account error for note with id {0}: {1}")]
-    AccountError(NoteId, AccountError),
-    #[error("asset error for note with id {0}: {1}")]
-    AssetError(NoteId, AssetError),
-    #[error("expected {1} note inputs for note with id {0}")]
-    WrongNumInputs(NoteId, usize),
-    #[error("note input representing block with value {1} for note with id {0}")]
-    BlockNumberError(NoteId, u64),
 }